Made energy class setup to depend on the micro sim class. Updated test. Moved exchange class functions to micro exch file
