+++ native/include/c_energy.h
@@ -1,5 +1,7 @@
+#pragma once
 #include<cmath>
 #include<iostream>
+#include "c_micro_sim.h"
 
 class Energy {
 public:
@@ -18,13 +20,11 @@ public:
     double *coordinates;
     int *ngbs;
     double compute_energy();
-    void setup(int nx, int ny, int nz, double dx, double dy, double dz,
-               double unit_length, double *spin, double *Ms, double *Ms_inv,
-               double *coordinates, int *ngbs, 
-               double *energy, double *field
-               );
+    // Will get the parameters from a simulation class
+    void _setup(MicroSim * sim);
     virtual void compute_field(double t) {};
 };
+
 
 class ExchangeEnergy : public Energy {
 public:
@@ -32,11 +32,27 @@ public:
         std::cout << "Instatiating ExchangeEnergy class; at " << this << "\n";
         this->interaction_id = 1;
     };
-    ~ExchangeEnergy() {std::cout << "Killing B\n";};
-    void init(double *A) {
-        this->set_up = false;
+    ~ExchangeEnergy() {std::cout << "Killing Exchange\n";};
+    void setup(double *A, MicroSim * sim) {
+        _setup(sim);
         this->A = A;
     }
     double *A;
     void compute_field(double t);
 };
+
+
+// class AnisotropyEnergy : public Energy {
+// public:
+//     AnisotropyEnergy() {
+//         std::cout << "Instatiating AnisotropyEnergy class; at " << this << "\n";
+//         this->interaction_id = 1;
+//     };
+//     ~AnisotropyEnergy() {std::cout << "Killing Anisotropy\n";};
+//     void init(double *Ku) {
+//         this->set_up = false;
+//         this->Ku = Ku;
+//     }
+//     double *A;
+//     void compute_field(double t);
+// };
