+++ native/include/c_micro_sim.h
@@ -1,3 +1,4 @@
+#pragma once
 #include<cmath>
 #include<iostream>
 #include<vector>
